set up routes & corresponding elements
