+++ src/pages/index.js
@@ -1,7 +1,19 @@
-export { HomePage } from "./home";
-export { LoginPage } from "./login";
-export { ProfilePage } from "./profile";
-export { RegisterPage } from "./register";
-export { IngredientPage } from "./ingredient";
-export { ResetPasswordPage } from "./reset-password";
-export { ForgotPasswordPage } from "./forgot-password";
+import HomePage from "./home/home";
+import LoginPage from "./login/login";
+import ProfilePage from "./profile/profile";
+import RegisterPage from "./register/register";
+import IngredientPage from "./ingredient/ingredient";
+import NotFoundPage from "./errors/not-found/not-found";
+import ResetPasswordPage from "./reset-password/reset-password";
+import ForgotPasswordPage from "./forgot-password/forgot-password";
+
+export {
+  HomePage,
+  LoginPage,
+  ProfilePage,
+  RegisterPage,
+  NotFoundPage,
+  IngredientPage,
+  ResetPasswordPage,
+  ForgotPasswordPage    
+};
